chore: improve component separation of concerns, hooks, purity
