+++ frontend/pages/customDomains/components/DomainLink.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+function DomainLink({ domain }) {
+  return (
+    <a
+      target="_blank"
+      rel="noopener noreferrer"
+      title={`Link to site's domain ${domain}`}
+      href={`https://${domain}`}
+    >
+      {domain}
+    </a>
+  );
+}
+
+DomainLink.propTypes = {
+  domain: PropTypes.string.isRequired,
+};
