+++ frontend/pages/organizations/Edit.jsx
@@ -2,18 +2,15 @@ import React, { useEffect, useReducer } from 'react';
 import { useParams } from 'react-router-dom';
 import { success } from 'react-notification-system-redux';
 import { useSelector } from 'react-redux';
-import PropTypes from 'prop-types';
 
 import federalistApi from '@util/federalistApi';
-import { timeFrom } from '@util/datetime';
 import { sandboxMsg } from '@util';
+
 import LoadingIndicator from '@shared/LoadingIndicator';
 import AlertBanner from '@shared/alertBanner';
 
 import AddUserForm from './components/AddUserForm';
-import RemoveUserForm from './components/RemoveUserForm';
-import ResendInviteForm from './components/ResendInviteForm';
-import UpdateUserForm from './components/UpdateUserForm';
+import OrganizationTable from './components/OrganizationTable';
 
 function successNotification(message) {
   return success({
@@ -22,11 +19,6 @@ function successNotification(message) {
     position: 'tr',
     autoDismiss: 3,
   });
-}
-
-function showRemoveConfirm(member, org) {
-  // eslint-disable-next-line no-alert
-  return window.confirm(`Are you sure you want to remove ${member.User.UAAIdentity.email} from ${org.name}?`);
 }
 
 function getInvitationSentMsg(mostRecentlyAddedUser) {
@@ -87,7 +79,7 @@ const initialState = {
   mostRecentlyAddedUser: null,
 };
 
-function Edit({ actions }) {
+function Edit() {
   const [state, dispatch] = useReducer(reducer, initialState);
   const currentUser = useSelector(({ user }) => user.data);
   const { id } = useParams();
@@ -95,9 +87,9 @@ function Edit({ actions }) {
   useEffect(() => {
     const fetchInitialData = async () => {
       const [members, org, roles] = await Promise.all([
-        actions.fetchOrganizationMembers(id),
-        actions.fetchOrganization(id),
-        actions.fetchRoles(),
+        federalistApi.fetchOrganizationMembers(id),
+        federalistApi.fetchOrganization(id),
+        federalistApi.fetchRoles(),
       ]);
       dispatch({ type: 'init', payload: { members, org, roles } });
     };
@@ -127,8 +119,6 @@ function Edit({ actions }) {
     .filter(member => member.User.id !== currentUser.id)
     .sort((a, b) => a.User.UAAIdentity.email > b.User.UAAIdentity.email);
 
-  const getSandboxMsg = days => <span>{sandboxMsg(days)}</span>;
-
   return (
     <div className="grid-col-12">
       <div className="page-header grid-row">
@@ -142,7 +132,7 @@ function Edit({ actions }) {
           && (
           <AlertBanner
             status="warning"
-            message={getSandboxMsg(org.daysUntilSandboxCleaning)}
+            message={sandboxMsg(org.daysUntilSandboxCleaning)}
             alertRole={false}
           />
           )}
@@ -160,7 +150,7 @@ function Edit({ actions }) {
           className="well"
           roleOptions={roleOptions}
           onSubmit={
-            data => actions.inviteToOrganization(org.id, data)
+            data => federalistApi.inviteToOrganization(org.id, data)
           }
           onSubmitSuccess={
             ({ member, invite: { email, link } }, reduxDispatch) => {
@@ -172,132 +162,18 @@ function Edit({ actions }) {
             }
           }
         />
-        <table className="usa-table usa-table--borderless usa-table--stacked log-table org-member-table width-full table-full-width">
-          <thead>
-            <tr>
-              <th scope="col">Email</th>
-              <th scope="col">Role</th>
-              <th scope="col">Added</th>
-              <th scope="col">Last Signed In</th>
-              <th scope="col">Actions</th>
-            </tr>
-          </thead>
-          <tbody>
-            {currentMember?.User && (
-              <tr key={currentMember.User?.id}>
-                <th scope="row" data-title="Email">{currentMember.User?.UAAIdentity?.email}</th>
-                <td data-title="Role">
-                  manager
-                </td>
-                <td data-title="Added">
-                  {timeFrom(currentMember?.createdAt)}
-                </td>
-                <td data-title="Last Signed In">
-                  {currentMember?.User?.signedInAt ? timeFrom(currentMember.User.signedInAt) : 'Never'}
-                </td>
-                <td label="Actions" data-title="Actions" className="table-actions" />
-              </tr>
-            )}
-
-            {sortedMembers.map(member => (
-              <tr key={member.User.id}>
-                <th scope="row" data-title="Email">{member.User.UAAIdentity.email}</th>
-                <td data-title="Role">
-                  <span className="usa-sr-only">Role</span>
-                  <UpdateUserForm
-                    form={`updateOrganizationUser-${member.User.id}`}
-                    initialValues={{ roleId: member.Role.id }}
-                    roleOptions={roleOptions}
-                    onSubmit={
-                      ({ roleId }) => actions.updateOrganizationRole(org.id, roleId, member.User.id)
-                    }
-                    onSubmitSuccess={
-                      (updatedMember, reduxDispatch) => {
-                        dispatch({ type: 'updateMember', payload: updatedMember });
-                        reduxDispatch(successNotification('Successfully updated user.'));
-                      }
-                    }
-                  />
-                </td>
-                <td data-title="Added">
-                  {timeFrom(member.createdAt)}
-                </td>
-                <td data-title="Last Signed In">
-                  {member.User.signedInAt ? timeFrom(member.User.signedInAt) : 'Never'}
-                </td>
-                <td data-title="Actions" className="table-actions">
-                  <span className="usa-sr-only">User actions</span>
-                  <RemoveUserForm
-                    form={`removeOrganizationUser-${member.User.id}`}
-                    onSubmit={() => true}
-                    onSubmitSuccess={
-                      async (_, reduxDispatch) => {
-                        if (!showRemoveConfirm(member, org)) {
-                          return;
-                        }
-
-                        await actions.removeOrganizationRole(org.id, member.User.id);
-                        dispatch({ type: 'removeMember', payload: member.User.id });
-                        reduxDispatch(successNotification('Successfully removed user.'));
-                      }
-                    }
-                  />
-                  { (member.User.UAAIdentity.origin === 'uaa' || member.User.UAAIdentity.origin === 'cloud.gov')
-                    && !member.User.signedInAt && (
-                    <ResendInviteForm
-                      form={`resendInvite-${member.User.id}`}
-                      onSubmit={() => true}
-                      onSubmitSuccess={
-                        async (_, reduxDispatch) => {
-                          await actions.inviteToOrganization(
-                            org.id,
-                            {
-                              uaaEmail: member.User.UAAIdentity.email,
-                              user: member.User.id,
-                              roleID: member.Role.id,
-                              isResend: true,
-                            }
-                          );
-                          reduxDispatch(successNotification('Successfully resent invitation.'));
-                        }
-                      }
-                    />
-                  )}
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
+        <OrganizationTable
+          org={org}
+          currentMember={currentMember}
+          sortedMembers={sortedMembers}
+          roleOptions={roleOptions}
+          dispatch={dispatch}
+          successNotification={successNotification}
+        />
       </div>
     </div>
   );
 }
 
-Edit.propTypes = {
-  actions: PropTypes.shape({
-    fetchOrganizationMembers: PropTypes.func.isRequired,
-    fetchOrganization: PropTypes.func.isRequired,
-    fetchRoles: PropTypes.func.isRequired,
-    inviteToOrganization: PropTypes.func.isRequired,
-    resendInviteToOrganization: PropTypes.func,
-    removeOrganizationRole: PropTypes.func.isRequired,
-    updateOrganizationRole: PropTypes.func.isRequired,
-  }).isRequired,
-};
-
-const WrappedEdit = props => (
-  <Edit
-    actions={{
-      fetchOrganizationMembers: federalistApi.fetchOrganizationMembers,
-      fetchOrganization: federalistApi.fetchOrganization,
-      fetchRoles: federalistApi.fetchRoles,
-      inviteToOrganization: federalistApi.inviteToOrganization,
-      removeOrganizationRole: federalistApi.removeOrganizationRole,
-      updateOrganizationRole: federalistApi.updateOrganizationRole,
-    }}
-    {...props}
-  />
-);
-
 export { Edit };
-export default WrappedEdit;
+export default Edit;
