+++ frontend/pages/sites/publishedFilesTable.jsx
@@ -6,99 +6,8 @@ import publishedFileActions from '@actions/publishedFileActions';
 import LoadingIndicator from '@shared/LoadingIndicator';
 import AlertBanner from '@shared/alertBanner';
 
-import globals from '../../globals';
-
-function renderPagingButtons(currentPage, lastPage, setCurrentPage) {
-  const shouldDisableNextPage = currentPage === lastPage;
-  const shouldDisablePreviousPage = currentPage === 0;
-  const shouldShowButtons = !(lastPage !== null && lastPage === 0);
-
-  function previousPage() {
-    setCurrentPage(currentPage - 1);
-  }
-
-  function nextPage() {
-    setCurrentPage(currentPage + 1);
-  }
-
-  const prevButtonClass = `${shouldDisablePreviousPage ? 'usa-button-disabled' : 'usa-button'}`;
-  const nextButtonClass = `pull-right ${shouldDisableNextPage ? 'usa-button-disabled' : 'usa-button'}`;
-
-  if (!shouldShowButtons) {
-    return null;
-  }
-
-  return (
-    <nav className="pagination" aria-label="Pagination">
-      <button
-        className={prevButtonClass}
-        disabled={shouldDisablePreviousPage}
-        onClick={previousPage}
-        title="View the previous page of published files"
-        type="button"
-      >
-        &laquo; Previous
-      </button>
-
-      <button
-        className={nextButtonClass}
-        disabled={shouldDisableNextPage}
-        onClick={nextPage}
-        title="View the next page of published files"
-        type="button"
-      >
-        Next &raquo;
-      </button>
-    </nav>
-  );
-}
-
-function renderBranchFileRow(file) {
-  let viewFileLink;
-  const branch = file.publishedBranch.name;
-  switch (branch) {
-    case file.publishedBranch.site.defaultBranch:
-      viewFileLink = `${file.publishedBranch.site.viewLink}${file.name}`;
-      break;
-    case file.publishedBranch.site.demoBranch:
-      viewFileLink = `${file.publishedBranch.site.demoViewLink}${file.name}`;
-      break;
-    default:
-      viewFileLink = `${file.publishedBranch.site.previewLink}${branch}/${file.name}`;
-  }
-  return (
-    <tr key={viewFileLink}>
-      <th scope="row">{file.name}</th>
-      <td><a href={viewFileLink} target="_blank" rel="noopener noreferrer">View</a></td>
-    </tr>
-  );
-}
-
-function renderPublishedFilesTable(files, name, currentPage, lastPage, setCurrentPage) {
-  return (
-    <div>
-      <h3>{name}</h3>
-      <p>
-        Use this page to audit the files that
-        {` ${globals.APP_NAME} `}
-        has publicly published.
-        Up to 200 files are shown per page.
-      </p>
-      <table className="usa-table usa-table--borderless width-full usa-table--stacked log-table table-full-width">
-        <thead>
-          <tr>
-            <th scope="col">File</th>
-            <th scope="col">Actions</th>
-          </tr>
-        </thead>
-        <tbody>
-          { files.filter(f => !!f.name).map(renderBranchFileRow) }
-        </tbody>
-      </table>
-      { renderPagingButtons(currentPage, lastPage, setCurrentPage) }
-    </div>
-  );
-}
+import PagingButtons from './components/PagingButtons';
+import FilesTable from './components/FilesTable';
 
 function PublishedFilesTable() {
   const { id, name } = useParams();
@@ -140,10 +49,18 @@ function PublishedFilesTable() {
       />
     );
   }
-  return renderPublishedFilesTable(
-    publishedFiles.data.files,
-    name, currentPage, lastPage,
-    setCurrentPage
+
+  return (
+    <FilesTable
+      files={publishedFiles.data.files}
+      name={name}
+    >
+      <PagingButtons
+        currentPage={currentPage}
+        lastPage={lastPage}
+        setCurrentPage={setCurrentPage}
+      />
+    </FilesTable>
   );
 }
